move tests from /model to /unit
