+++ test/unit/sequence_result_test.rb
