+++ test/unit/server_capabilities_test.rb
