+++ test/unit/test_result_test.rb
